add hook useSongs
